+++ app/gatherDemographics.py
@@ -104,8 +104,7 @@ def get_demo():
     for acq in session_container.acquisitions.iter():
         # print(acq.label)
         acq = acq.reload()
-        if 'T2' in acq.label and 'AXI' in acq.label and 'Segmentation' not in acq.label or 't1_mprage_sag_p2_iso' in acq.label: 
-            # PRIMES only has these scan names: and 'NOT FOR DIAGNOSTIC USE' not in acq.label
+        if 'T2' in acq.label and 'AXI' in acq.label and 'Segmentation' not in acq.label: 
             for file_obj in acq.files: # get the files in the acquisition
                 # Screen file object information & download the desired file
                 if file_obj['type'] == 'dicom':