+++ start.sh
@@ -1,6 +1,6 @@
 #!/bin/env bash 
 . $FSLDIR/etc/fslconf/fsl.sh
 echo "FSLOUTPUTTYPE set to $FSLOUTPUTTYPE"
-python3 /flywheel/v0/run.py
+python3 -u /flywheel/v0/run.py